+++ firebase-vscode/src/cli.ts
@@ -1,6 +1,3 @@
-import * as vscode from "vscode";
-import { inspect } from "util";
-
 import {
   getAllAccounts,
   getGlobalDefaultAccount,
@@ -10,13 +7,12 @@ import {
 import { logoutAction } from "../../src/commands/logout";
 import { listFirebaseProjects } from "../../src/management/projects";
 import { requireAuth } from "../../src/requireAuth";
-import { Account, User } from "../../src/types/auth";
+import { Account, Tokens, User } from "../../src/types/auth";
 import { Options } from "../../src/options";
 import { currentOptions, getCommandOptions } from "./options";
-import { ServiceAccount } from "../common/types";
 import { EmulatorUiSelections } from "../common/messaging/types";
 import { pluginLogger } from "./logger-wrapper";
-import { setAccessToken } from "../../src/apiv2";
+import { getAccessToken, setAccessToken } from "../../src/apiv2";
 import {
   startAll as startAllEmulators,
   cleanShutdown as stopAllEmulators,
@@ -32,65 +28,45 @@ import * as commandUtils from "../../src/emulator/commandUtils";
 import { currentUser } from "./core/user";
 import { firstWhere } from "./utils/signal";
 export { Emulators };
-/**
- * Try to get a service account by calling requireAuth() without
- * providing any account info.
- */
-async function getServiceAccount() {
-  let email = null;
-  try {
-    // Make sure no user/token is sent
-    // to requireAuth which would prevent autoAuth() from being reached.
-    // We do need to send isVSCE to prevent project selection popup
-    const optionsMinusUser = await getCommandOptions(undefined, {
-      ...currentOptions.value,
-    });
-    delete optionsMinusUser.user;
-    delete optionsMinusUser.tokens;
-    delete optionsMinusUser.token;
-    email = (await requireAuth(optionsMinusUser)) || null;
-  } catch (e) {
-    let errorMessage = e.message;
-    if (e.original?.message) {
-      errorMessage += ` (original: ${e.original.message})`;
-    }
-    pluginLogger.debug(
-      `No service account found (this may be normal), ` +
-        `requireAuth error output: ${errorMessage}`,
-    );
-    return null;
-  }
-  if (process.env.WORKSPACE_SERVICE_ACCOUNT_EMAIL) {
-    // If Monospace, get service account email using env variable as
-    // the metadata server doesn't currently return the credentials
-    // for the workspace service account. Remove when Monospace is
-    // updated to return credentials through the metadata server.
-    pluginLogger.debug(
-      `Using WORKSPACE_SERVICE_ACCOUNT_EMAIL env ` +
-        `variable to get service account email: ` +
-        `${process.env.WORKSPACE_SERVICE_ACCOUNT_EMAIL}`,
-    );
-    return process.env.WORKSPACE_SERVICE_ACCOUNT_EMAIL;
-  }
-  pluginLogger.debug(
-    `Got service account email through credentials:` + ` ${email}`,
-  );
-  return email;
-}
 
 /**
  * Wrap the CLI's requireAuth() which is normally run before every command
  * requiring user to be logged in. The CLI automatically supplies it with
  * account info if found in configstore so we need to fill that part in.
+ *
  */
-async function requireAuthWrapper(showError: boolean = true): Promise<boolean> {
-  // Try to get global default from configstore. For some reason this is
+export async function requireAuthWrapper(
+  showError: boolean = true,
+): Promise<User | null> {
+  // Try to get global default from configstore
   pluginLogger.debug("requireAuthWrapper");
   let account = getGlobalDefaultAccount();
   // often overwritten when restarting the extension.
+  const accounts = getAllAccounts();
+
+  // helper to determine if VSCode options has the same account as global default
+  function isUserMatching(account: Account, options: Options) {
+
+    if (!options.user || !options.tokens) {
+      return false;
+    }
+
+    const optionsUser = options.user as User;
+    const optionsTokens = options.tokens as Tokens;
+    return (
+      account &&
+      account.user.email === optionsUser.email &&
+      account.tokens.refresh_token === optionsTokens.refresh_token // Should check refresh token which is consistent, not access_token which is short lived. 
+    );
+  }
+
+  // only add account options when vscode is missing account information
+  if (!isUserMatching(account, currentOptions.value)) {
+    currentOptions.value = { ...currentOptions.value, ...account };
+  }
+
   if (!account) {
     // If nothing in configstore top level, grab the first "additionalAccount"
-    const accounts = getAllAccounts();
     for (const additionalAccount of accounts) {
       if (additionalAccount.user.email === currentUser.value.email) {
         account = additionalAccount;
@@ -98,50 +74,33 @@ async function requireAuthWrapper(showError: boolean = true): Promise<boolean> {
       }
     }
   }
-  const commandOptions = await getCommandOptions(undefined, {
-    ...currentOptions.value,
-  });
   // `requireAuth()` is not just a check, but will also register SERVICE
   // ACCOUNT tokens in memory as a variable in apiv2.ts, which is needed
   // for subsequent API calls. Warning: this variable takes precedence
   // over Google login tokens and must be removed if a Google
   // account is the current user.
   try {
-    const serviceAccountEmail = await getServiceAccount();
-    // Priority 1: Service account exists and is the current selected user
-    if (
-      serviceAccountEmail &&
-      currentUser.value.email === serviceAccountEmail
-    ) {
-      // requireAuth should have been run and apiv2 token should be stored
-      // already due to getServiceAccount() call above.
-      return true;
-    } else if (account) {
-      // Priority 2: Google login account exists and is the currently selected
-      // user
-      // Priority 3: Google login account exists and there is no selected user
-      // Clear service account access token from memory in apiv2.
-      setAccessToken();
-      await requireAuth({ ...commandOptions, ...account });
-      return true;
-    } else if (serviceAccountEmail) {
-      // Priority 4: There is a service account but it's not set as
-      // currentUser for some reason, but there also isn't an oauth account.
-      // requireAuth was already run as part of getServiceAccount() above
-      return true;
+    const userEmail = await requireAuth(currentOptions.value); // client email
+    // SetAccessToken is necessary here to ensure that access_tokens are available when:
+    // - we are using tokens from configstore (aka those set by firebase login), AND
+    // - we are calling CLI code that skips Command (where we normally call this)
+
+    setAccessToken(await getAccessToken()); 
+    if (userEmail) {
+      pluginLogger.debug("User found: ", userEmail);
+
+      // VSCode only has the concept of a single user
+      return getGlobalDefaultAccount().user;
     }
+
     pluginLogger.debug("No user found (this may be normal)");
-    return false;
+    return null;
   } catch (e) {
     if (showError) {
       // Show error to user - show a popup and log it with log level "error".
       pluginLogger.error(
         `requireAuth error: ${e.original?.message || e.message}`,
       );
-      vscode.window.showErrorMessage("Not logged in", {
-        modal: true,
-        detail: `Log in by clicking "Sign in with Google" in the sidebar.`,
-      });
     } else {
       // User shouldn't need to see this error - not actionable,
       // but we should log it for debugging purposes.
@@ -150,23 +109,8 @@ async function requireAuthWrapper(showError: boolean = true): Promise<boolean> {
         e.original?.message || e.message,
       );
     }
-    return false;
+    return null;
   }
-}
-
-export async function getAccounts(): Promise<Array<Account | ServiceAccount>> {
-  // Get Firebase login accounts
-  const accounts: Array<Account | ServiceAccount> = getAllAccounts();
-  pluginLogger.debug(`Found ${accounts.length} non-service accounts.`);
-  // Get other accounts (assuming service account for now, could also be glogin)
-  const serviceAccountEmail = await getServiceAccount();
-  if (serviceAccountEmail) {
-    pluginLogger.debug(`Found service account: ${serviceAccountEmail}`);
-    accounts.push({
-      user: { email: serviceAccountEmail, type: "service_account" },
-    });
-  }
-  return accounts;
 }
 
 export async function logoutUser(email: string): Promise<void> {
@@ -183,8 +127,8 @@ export async function login() {
 }
 
 export async function listProjects() {
-  const loggedIn = await requireAuthWrapper(false);
-  if (!loggedIn) {
+  const loggedInUser = await requireAuthWrapper(false);
+  if (!loggedInUser) {
     return [];
   }
   return listFirebaseProjects();
@@ -193,9 +137,10 @@ export async function listProjects() {
 export async function emulatorsStart(
   emulatorUiSelections: EmulatorUiSelections,
 ) {
-  const only = emulatorUiSelections.mode === "dataconnect"
-        ? `${Emulators.DATACONNECT},${Emulators.AUTH}`
-        : "";
+  const only =
+    emulatorUiSelections.mode === "dataconnect"
+      ? `${Emulators.DATACONNECT},${Emulators.AUTH}`
+      : "";
   const commandOptions = await getCommandOptions(undefined, {
     ...(await firstWhere(
       // TODO use firstWhereDefined once currentOptions are undefined if not initialized yet