+++ firebase-vscode/src/data-connect/emulator-stream.ts
@@ -10,6 +10,7 @@ enum Kind {
   SQL_CONNECTION = "SQL_CONNECTION",
   SQL_MIGRATION = "SQL_MIGRATION",
   VERTEX_AI = "VERTEX_AI",
+  FILE_RELOAD = "FILE_RELOAD",
 }
 enum Severity {
   SEVERITY_UNSPECIFIED = "SEVERITY_UNSPECIFIED",
@@ -36,13 +37,14 @@ export async function runEmulatorIssuesStream(
   configs: ResolvedDataConnectConfigs,
   fdcEndpoint: string,
   isPostgresEnabled: Signal<boolean>,
+  schemaReloadFunc: () => void,
 ) {
   const obsErrors = await getEmulatorIssuesStream(configs, fdcEndpoint);
   const obsConverter = {
     next(nextResponse: EmulatorIssueResponse) {
       if (nextResponse.result?.issues?.length) {
         for (const issue of nextResponse.result.issues) {
-          displayAndHandleIssue(issue, isPostgresEnabled);
+          displayAndHandleIssue(issue, isPostgresEnabled, schemaReloadFunc);
         }
       }
     },
@@ -59,18 +61,23 @@ export async function runEmulatorIssuesStream(
 /**
  * Based on the severity of the issue, either log, display notification, or display interactive popup to the user
  */
-export function displayAndHandleIssue(issue: EmulatorIssue, isPostgresEnabled: Signal<boolean>) {
+export function displayAndHandleIssue(
+  issue: EmulatorIssue,
+  isPostgresEnabled: Signal<boolean>,
+  schemaReloadFunc: () => void,
+) {
   const issueMessage = `Data Connect Emulator: ${issue.kind.toString()} - ${issue.message}`;
   if (issue.severity === Severity.ALERT) {
     vscode.window.showErrorMessage(issueMessage);
-  } else if (issue.severity === Severity.NOTICE) {
-    vscode.window.showWarningMessage(issueMessage);
   }
   emulatorOutputChannel.appendLine(issueMessage);
 
   // special handlings
   if (issue.kind === Kind.SQL_CONNECTION) {
     isPostgresEnabled.value = false;
+  }
+  if (issue.kind === Kind.FILE_RELOAD) {
+    schemaReloadFunc();
   }
 }